Fix test and small optimization to NodeAssigner before moving branch up.

+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/dht/impl/ActiveDHTNodeController.java
@@ -77,7 +77,9 @@ class ActiveDHTNodeController extends AbstractDHTController {
         sendUpdatedCapabilities();
         
         try {
-            if(DHTSettings.PERSIST_DHT.getValue()) {
+        //TODO: fix this
+            if(DHTSettings.PERSIST_DHT.getValue() 
+                    && FILE.exists() && FILE.isFile()) {
                 FileOutputStream out = new FileOutputStream(FILE);
                 dht.store(out);
                 out.close();
